Move more aecdata classes out of __init__.py
